v 1.1 release: * fix fixpipe while writing to dead socket * add try for reconnection to previously dead fd on running * add pdu validation output * formatting, cleanup

+++ src/tui.c
@@ -212,8 +212,8 @@ init_tui(global_t *globals) {
     refresh();
 
     // init header
-    u8 header_bottom = 16;
-    wheader          = newwin(header_bottom, COLS, 0, 0);
+    int header_bottom = 16;
+    wheader           = newwin(header_bottom, COLS, 0, 0);
     redraw_header();
 
     // init logd
